+++ sermant-plugins/sermant-tag-transmission/tag-transmission-plugin/src/main/java/com/huaweicloud/sermant/tag/transmission/interceptors/RocketmqProducerInterceptor.java
@@ -89,13 +89,6 @@ public class RocketmqProducerInterceptor extends AbstractClientInterceptor {
 
     @Override
     public ExecuteContext doAfter(ExecuteContext context) {
-        TrafficUtils.removeTrafficTag();
-        return context;
-    }
-
-    @Override
-    public ExecuteContext onThrow(ExecuteContext context) {
-        TrafficUtils.removeTrafficTag();
         return context;
     }
 }
