+++ sermant-plugins/sermant-tag-transmission/tag-transmission-plugin/src/test/java/com/huaweicloud/sermant/tag/transmission/interceptors/RocketmqConsumerInterceptorInterceptorTest.java
@@ -19,10 +19,15 @@ package com.huaweicloud.sermant.tag.transmission.interceptors;
 import com.huaweicloud.sermant.core.plugin.agent.entity.ExecuteContext;
 import com.huaweicloud.sermant.core.utils.tag.TrafficUtils;
 
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.common.header.Headers;
+import org.apache.rocketmq.common.message.Message;
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mockito;
+import org.powermock.api.mockito.PowerMockito;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
+import org.powermock.reflect.Whitebox;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -31,20 +36,48 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * 测试KafkaConsumerRecordInterceptor
+ * 测试RocketmqConsumerInterceptor
  *
  * @author tangle
  * @since 2023-07-27
  */
-public class KafkaConsumerRecordInterceptorTest extends BaseTest {
-    private final KafkaConsumerRecordInterceptor interceptor;
+@RunWith(PowerMockRunner.class)
+@PrepareForTest(RocketmqConsumerInterceptor.class)
+public class RocketmqConsumerInterceptorInterceptorTest extends BaseInterceptorTest {
+    private final RocketmqConsumerInterceptor interceptor;
 
-    public KafkaConsumerRecordInterceptorTest() {
-        interceptor = new KafkaConsumerRecordInterceptor();
+    public RocketmqConsumerInterceptorInterceptorTest() throws Exception {
+        RocketmqConsumerInterceptor interceptorBase = new RocketmqConsumerInterceptor();
+        interceptor = PowerMockito.spy(interceptorBase);
+        PowerMockito.doReturn(true).when(interceptor, "isRocketMqStackTrace", Mockito.any());
     }
 
     @Test
-    public void testKafkaConsumer() {
+    public void testIsAvailable() throws Exception {
+        StackTraceElement[] stackTraceElements = new StackTraceElement[2];
+        RocketmqConsumerInterceptor interceptorBase = new RocketmqConsumerInterceptor();
+
+        // 测试可通过
+        stackTraceElements[0] = new StackTraceElement("org.apache.rocketmq.common.message.Message", "xxxMethod",
+                "xxxFile", 1);
+        stackTraceElements[1] = new StackTraceElement("org.apache.xxx", "xxxMethod",
+                "xxxFile", 1);
+        boolean resultAccess = Whitebox.invokeMethod(interceptorBase, "isRocketMqStackTrace",
+                (Object) stackTraceElements);
+        Assert.assertTrue(resultAccess);
+
+        // 测试不可通过
+        stackTraceElements[0] = new StackTraceElement("org.apache.rocketmq.common.message.Message", "xxxMethod",
+                "xxxFile", 1);
+        stackTraceElements[1] = new StackTraceElement("org.apache.rocketmq", "xxxMethod",
+                "xxxFile", 1);
+        boolean resultRefuse = Whitebox.invokeMethod(interceptorBase, "isRocketMqStackTrace",
+                (Object) stackTraceElements);
+        Assert.assertFalse(resultRefuse);
+    }
+
+    @Test
+    public void testRocketmqConsumer() {
         ExecuteContext context;
         Map<String, List<String>> addHeaders = new HashMap<>();
         Map<String, List<String>> tags = new HashMap<>();
@@ -64,13 +97,10 @@ public class KafkaConsumerRecordInterceptorTest extends BaseTest {
         Assert.assertNull(TrafficUtils.getTrafficTag().getTag().get("name"));
 
         // 有Headers有Tags
-        List<String> ids = new ArrayList<>();
-        ids.add("testId001");
-        ids.add("testId002");
-        tags.put("id", ids);
+        tags.put("id", Collections.singletonList("testId001"));
         context = buildContext(addHeaders, tags);
         interceptor.before(context);
-        Assert.assertEquals(TrafficUtils.getTrafficTag().getTag().get("id").size(), 2);
+        Assert.assertEquals(TrafficUtils.getTrafficTag().getTag().get("id").get(0), "testId001");
         Assert.assertNull(TrafficUtils.getTrafficTag().getTag().get("name"));
 
         // 第二次消费，测试tags是否污染其他消费
@@ -81,7 +111,7 @@ public class KafkaConsumerRecordInterceptorTest extends BaseTest {
         context = buildContext(addHeaders, tags);
         interceptor.before(context);
         Assert.assertNull(TrafficUtils.getTrafficTag().getTag().get("id"));
-        Assert.assertEquals(TrafficUtils.getTrafficTag().getTag().get("name").size(), 1);
+        Assert.assertEquals(TrafficUtils.getTrafficTag().getTag().get("name").get(0), "testName001");
 
         // 测试TagTransmissionConfig开关关闭时
         TrafficUtils.removeTrafficTag();
@@ -91,20 +121,19 @@ public class KafkaConsumerRecordInterceptorTest extends BaseTest {
     }
 
     private ExecuteContext buildContext(Map<String, List<String>> addHeaders, Map<String, List<String>> tags) {
-        ConsumerRecord consumerRecord = new ConsumerRecord("topic", 1, 1L, "key", "value");
-        Headers headers = consumerRecord.headers();
+        Message message = new Message();
         for (Map.Entry<String, List<String>> entry : addHeaders.entrySet()) {
             for (String val : entry.getValue()) {
-                headers.add(entry.getKey(), val.getBytes());
+                message.putUserProperty(entry.getKey(), val);
             }
         }
         for (Map.Entry<String, List<String>> tag : tags.entrySet()) {
             for (String val : tag.getValue()) {
-                headers.add(tag.getKey(), val.getBytes());
+                message.putUserProperty(tag.getKey(), val);
             }
         }
         return ExecuteContext.forMemberMethod(
-                consumerRecord,
+                message,
                 null,
                 null,
                 null,
